requirements.txt
beartype==0.14.1
GitPython==3.1.31
hydra-core==1.3.2
loguru==0.7.0
pyshortcuts==1.9.0
python-dotenv==1.0.0
python-gitlab==3.15.0
rich==13.4.1
selenium==4.17.2
tqdm==4.56.0
webdriver-manager==4.0.1
